requirements.txt
asgiref==3.5.1
asttokens==2.0.5
attrs==21.4.0
backcall==0.2.0
backports.zoneinfo==0.2.1
decorator==5.1.1
Django==4.0.4
django-cors-headers==3.12.0
django-extensions==3.1.5
django-model-utils==1.1.0
django-shell-plus==1.1.7
djangorestframework==3.13.1
executing==0.8.3
factory-boy==3.2.1
Faker==13.11.1
iniconfig==1.1.1
ipdb==0.13.9
ipython==8.3.0
jedi==0.18.1
matplotlib-inline==0.1.3
packaging==21.3
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
pluggy==1.0.0
prompt-toolkit==3.0.29
psycopg2-binary==2.8.6
ptyprocess==0.7.0
pure-eval==0.2.2
py==1.11.0
Pygments==2.12.0
pyparsing==3.0.9
pytest==7.1.2
pytest-django==4.5.2
python-dateutil==2.8.2
pytz==2022.1
six==1.16.0
sqlparse==0.4.2
stack-data==0.2.0
toml==0.10.2
tomli==2.0.1
traitlets==5.2.0
wcwidth==0.2.5
